+++ src/components/Card.vue
@@ -1,7 +1,7 @@
 <template>
 <div class="card-pf" :class="{'card-pf-accented': accented}">
   <div v-if="showHeader" :class="{'card-pf-heading': showTitlesSeparator, 'card-pf-heading-no-bottom': !showTitlesSeparator}">
-    <pf-dropdown v-if="showFilterInHeader" class="card-pf-time-frame-filter pull-right" :text="currentFilter.label">
+    <pf-dropdown v-if="showFilterInHeader" class="card-pf-time-frame-filter" :text="currentFilter.label" menu-right>
       <li
         v-for="(item, i) in filter.filters"
         :key="i"
@@ -22,7 +22,7 @@
   </div>
 
   <div v-if="showFooter || showFilterInFooter" class="card-pf-footer">
-    <pf-dropdown v-if="showFilterInFooter" class="card-pf-time-frame-filter pull-right" :text="currentFilter.label">
+    <pf-dropdown v-if="showFilterInFooter" class="card-pf-time-frame-filter" :text="currentFilter.label" menu-right>
       <li
         v-for="(item, i) in filter.filters"
         :key="i"